# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OneAquaHealth Insight API",
    "version": "1.0.0",
    "description": "Read-only API over the frozen OneAquaHealth snapshot (pipeline Stage 8). Every number was computed offline by the pipeline; the API only filters and reshapes. W (water quality, higher = cleaner) and H (contamination hazard, higher = riskier) are relative 0-100 indices anchored to citizen rubrics, not lab measurements, and always travel with a 90% interval (q05-q95) and an evidence flag. Based on volunteer observations, not laboratory tests. Not an official health advisory."
  },
  "servers": [{ "url": "/" }],
...
  "paths": {
...
    "/api/v1/meta": {
      "get": {
        "tags": ["meta"],
        "summary": "Snapshot manifest, toggles, thresholds, model health, provenance and table glossary",
        "responses": {
          "200": { "$ref": "#/components/responses/Meta" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "503": { "$ref": "#/components/responses/NoSnapshot" }
        }
      }
    },
    "/api/v1/model/diagnostics": {
      "get": {
        "tags": ["meta"],
        "summary": "Convergence and posterior-predictive checks per model variant",
        "responses": {
          "200": { "$ref": "#/components/responses/Diagnostics" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "503": { "$ref": "#/components/responses/NoSnapshot" }
        }
      }
    },
    "/api/v1/model/params": {
      "get": {
        "tags": ["meta"],
        "summary": "Posterior summaries of model parameters",
        "parameters": [{ "$ref": "#/components/parameters/Variant" }],
        "responses": {
          "200": { "$ref": "#/components/responses/Params" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "503": { "$ref": "#/components/responses/NoSnapshot" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "Variant": {
        "name": "variant",
        "in": "query",
        "description": "Rain-assumption toggle: M1 = primary model with rain (default), M1_norain = model without rain. Allowed values are listed in /api/v1/meta.",
        "schema": { "type": "string", "enum": ["M1", "M1_norain"], "default": "M1" }
      },
...
    },
    "responses": {
      "BadRequest": { "description": "Invalid, unknown or repeated parameter", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "NoSnapshot": { "description": "No snapshot loaded yet", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Meta": { "description": "Snapshot metadata", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MetaEnvelope" } } } },
      "Diagnostics": { "description": "Diagnostics per variant", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListEnvelope" } } } },
      "Params": { "description": "Parameter summaries", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListEnvelope" } } } },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["status", "message"],
            "properties": { "status": { "type": "integer" }, "message": { "type": "string" }, "details": { "type": "object" } }
          }
        }
      },
...
      "Envelope": {
        "type": "object",
        "required": ["snapshot", "disclaimer", "data"],
        "properties": {
          "snapshot": {
            "type": "object",
            "properties": {
              "id": { "type": "string", "description": "Content address of the snapshot; also sent as the X-OAH-Snapshot header" },
              "schema_version": { "type": "string" },
              "latest_week": { "type": "string", "format": "date" },
              "created_at_utc": { "type": "string" }
            }
          },
          "disclaimer": { "type": "string" },
          "params": { "type": "object", "description": "The parameters used, after defaults" },
          "note": { "type": "string", "description": "What the toggles do and do not change for this endpoint" },
          "data": {}
        }
      },
      "ListEnvelope": {
        "allOf": [{ "$ref": "#/components/schemas/Envelope" }, { "properties": { "count": { "type": "integer" }, "data": { "type": "array" } } }]
      },
      "MetaEnvelope": {
        "allOf": [
          { "$ref": "#/components/schemas/Envelope" },
          {
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "manifest": { "type": "object" },
                  "weeks": { "type": "object", "properties": { "first": { "type": "string" }, "latest": { "type": "string" }, "count": { "type": "integer" } } },
                  "n_sites": { "type": "integer" },
                  "variants": { "type": "array", "items": { "type": "object", "properties": { "id": { "type": "string" }, "role": { "type": "string" }, "rain_assumed": { "type": "boolean" }, "description": { "type": "string" } } } },
                  "default_variant": { "type": "string" },
                  "sensitivity_levels": { "type": "array", "items": { "type": "object", "properties": { "id": { "type": "string" }, "delta_W": { "type": "number" } } } },
                  "default_sensitivity": { "type": "string" },
                  "recent_since": { "type": "string", "format": "date" },
                  "thresholds": { "type": "object" },
                  "model_health": { "type": "object", "additionalProperties": { "type": "boolean" } },
                  "languages": { "type": "array", "items": { "type": "string" } },
                  "tables": { "type": "array", "items": { "type": "object", "properties": { "name": { "type": "string" }, "comment": { "type": "string" } } } },
                  "views": { "type": "array", "items": { "type": "string" } },
                  "inputs": { "type": "array", "items": { "type": "object", "properties": { "label": { "type": "string" }, "path": { "type": "string" }, "sha256": { "type": "string" }, "n_rows": { "type": ["number", "null"] } } } }
                }
              }
            }
          }
        ]
      },
...
    }
  }
}